moved browser settings to new menu #426
