+++ src/mainAppBar/userProfile.component.test.tsx
@@ -9,7 +9,6 @@ import { push } from 'connected-react-router';
 import { Avatar } from '@material-ui/core';
 import thunk from 'redux-thunk';
 import TestAuthProvider from '../authentication/testAuthProvider';
-import { loadDarkModePreference } from '../state/actions/scigateway.actions';
 
 describe('User profile component', () => {
   let shallow;
@@ -95,22 +94,6 @@ describe('User profile component', () => {
     expect(wrapper.find('#simple-menu').first().prop('open')).toBeTruthy();
   });
 
-  it('opens cookie policy/management page if manage cookies clicked', () => {
-    const testStore = mockStore(state);
-    const wrapper = mount(
-      <Provider store={testStore}>
-        <UserProfileComponent />
-      </Provider>
-    );
-
-    // Click the user menu button and click on the manage cookies menu item.
-    wrapper.find('button').simulate('click');
-    wrapper.find('#item-manage-cookies').first().simulate('click');
-
-    expect(testStore.getActions().length).toEqual(1);
-    expect(testStore.getActions()[0]).toEqual(push('/cookies'));
-  });
-
   it('signs out if sign out clicked', () => {
     const testStore = mockStore(state);
     const wrapper = mount(
@@ -126,21 +109,5 @@ describe('User profile component', () => {
     expect(testStore.getActions().length).toEqual(2);
     expect(testStore.getActions()[0]).toEqual({ type: 'scigateway:signout' });
     expect(testStore.getActions()[1]).toEqual(push('/'));
-  });
-
-  it('sends load dark mode prefrence action if toggle dark mode is clicked', () => {
-    const testStore = mockStore(state);
-    const wrapper = mount(
-      <Provider store={testStore}>
-        <UserProfileComponent />
-      </Provider>
-    );
-
-    // Click the user menu button and click on the manage cookies menu item.
-    wrapper.find('button').simulate('click');
-    wrapper.find('#item-dark-mode').first().simulate('click');
-
-    expect(testStore.getActions().length).toEqual(1);
-    expect(testStore.getActions()[0]).toEqual(loadDarkModePreference(true));
   });
 });
