+++ src/mainAppBar/userProfile.component.tsx
@@ -1,9 +1,7 @@
 import React, { useState } from 'react';
 import { Dispatch, Action, AnyAction } from 'redux';
 import AccountCircleIcon from '@material-ui/icons/AccountCircle';
-import SettingsIcon from '@material-ui/icons/Settings';
 import LogoutIcon from '@material-ui/icons/ExitToApp';
-import BrightnessIcon from '@material-ui/icons/Brightness4';
 import {
   IconButton,
   withStyles,
@@ -22,10 +20,7 @@ import {
 import { StyleRules } from '@material-ui/core/styles';
 import { StateType, User } from '../state/state.types';
 import { getAppStrings, getString } from '../state/strings';
-import {
-  signOut,
-  loadDarkModePreference,
-} from '../state/actions/scigateway.actions';
+import { signOut } from '../state/actions/scigateway.actions';
 import { connect } from 'react-redux';
 import { AppStrings } from '../state/scigateway.types';
 import { ThunkDispatch } from 'redux-thunk';
@@ -37,14 +32,11 @@ interface UserProfileProps {
   loggedIn: boolean;
   user: User;
   res: AppStrings | undefined;
-  darkMode: boolean;
 }
 
 interface UserProfileDispatchProps {
   signIn: () => Action;
   signOut: () => void;
-  manageCookies: () => Action;
-  toggleDarkMode: (preference: boolean) => Action;
 }
 
 const styles = (theme: Theme): StyleRules =>
@@ -83,15 +75,6 @@ const UserProfileComponent = (
     closeMenu();
     props.signOut();
   };
-  const manageCookies = (): void => {
-    closeMenu();
-    props.manageCookies();
-  };
-  const toggleDarkMode = (): void => {
-    const toggledPreference = !props.darkMode;
-    localStorage.setItem('darkMode', toggledPreference.toString());
-    props.toggleDarkMode(toggledPreference);
-  };
   return (
     <div className="tour-user-profile">
       {props.loggedIn ? (
@@ -125,23 +108,6 @@ const UserProfileComponent = (
                 {props.user.username}
               </Typography>
             </div>
-            <Divider />
-            <MenuItem id="item-manage-cookies" onClick={manageCookies}>
-              <ListItemIcon>
-                <SettingsIcon />
-              </ListItemIcon>
-              <ListItemText
-                primary={getString(props.res, 'manage-cookies-button')}
-              />
-            </MenuItem>
-            <MenuItem id="item-dark-mode" onClick={toggleDarkMode}>
-              <ListItemIcon>
-                <BrightnessIcon />
-              </ListItemIcon>
-              <ListItemText
-                primary={getString(props.res, 'toggle-dark-mode')}
-              />
-            </MenuItem>
             <Divider />
             <MenuItem id="item-sign-out" onClick={logout}>
               <ListItemIcon>
@@ -182,7 +148,6 @@ const mapStateToProps = (state: StateType): UserProfileProps => ({
   user:
     state.scigateway.authorisation.provider.user || new UserInfo('anonymous'),
   res: getAppStrings(state, 'login'),
-  darkMode: state.scigateway.darkMode,
 });
 
 const mapDispatchToProps = (dispatch: Dispatch): UserProfileDispatchProps => ({
@@ -191,9 +156,6 @@ const mapDispatchToProps = (dispatch: Dispatch): UserProfileDispatchProps => ({
     const thunkDispatch = dispatch as ThunkDispatch<StateType, null, AnyAction>;
     thunkDispatch(signOut());
   },
-  manageCookies: () => dispatch(push('/cookies')),
-  toggleDarkMode: (preference: boolean) =>
-    dispatch(loadDarkModePreference(preference)),
 });
 
 export default connect(
